clean code: improves Pipelines
