+++ src/render_engine/TextPipeline.cpp
@@ -1,7 +1,7 @@
 #include "render_engine/TextPipeline.h"
 #include "CoreGraphicsContext.h"
 #include "render_engine/DescriptorSet.h"
-#include "render_engine/GraphicsPipeline.h"
+#include "render_engine/GeometryPipeline.h"
 #include "render_engine/RenderableGeometry.h"
 #include "render_engine/Sampler.h"
 #include "render_engine/Texture.h"
@@ -18,39 +18,37 @@ TextPipeline::TextPipeline(Window &window, std::shared_ptr<CoreGraphicsContext> 
                            std::vector<UniformBuffer> &uniform_buffers,
                            VkDescriptorSetLayout &descriptor_set_layout, Sampler &sampler,
                            Texture &texture)
-    : ctx(ctx) {
+    : m_ctx(ctx), m_pipeline(create_pipeline(descriptor_set_layout, swap_chain_manager)),
+      m_descriptor_pool(DescriptorPool(m_ctx, MAX_FRAMES_IN_FLIGHT)) {
 
     auto [graphicsQueue, presentQueue] = ctx->get_device_queues();
+    auto descriptor_set =
+        DescriptorSet(ctx, descriptor_set_layout, m_descriptor_pool.m_descriptor_pool,
+                      MAX_FRAMES_IN_FLIGHT, &uniform_buffers, &texture, &sampler);
+    geometry = std::make_unique<Geometry::Rectangle>(
+        ctx, swap_chain_manager, graphicsQueue, std::move(descriptor_set));
+}
 
+TextPipeline::~TextPipeline() {}
+
+Pipeline TextPipeline::create_pipeline(VkDescriptorSetLayout &descriptor_set_layout,
+                                       SwapChainManager &swap_chain_manager) {
     auto &resoure_manager = ResourceManager::get_instance();
     auto vert_shader_code = resoure_manager.get_resource<ShaderResource>("Vert");
     auto frag_shader_code = resoure_manager.get_resource<ShaderResource>("TextFragment");
 
     VkShaderModule vert_shader_module = createShaderModule(
-        ctx->device, vert_shader_code->bytes(), vert_shader_code->length());
+        m_ctx->device, vert_shader_code->bytes(), vert_shader_code->length());
+
     VkShaderModule frag_shader_module = createShaderModule(
-        ctx->device, frag_shader_code->bytes(), frag_shader_code->length());
+        m_ctx->device, frag_shader_code->bytes(), frag_shader_code->length());
 
-    graphics_pipeline = createGraphicsPipeline(ctx->device, vert_shader_module,
-                                               frag_shader_module, descriptor_set_layout,
-                                               pipeline_layout, swap_chain_manager);
+    Pipeline pipeline = Pipeline(m_ctx, descriptor_set_layout, vert_shader_module,
+                                 frag_shader_module, swap_chain_manager);
 
-    vkDestroyShaderModule(ctx->device, vert_shader_module, nullptr);
-    vkDestroyShaderModule(ctx->device, frag_shader_module, nullptr);
-
-    descriptor_pool = createDescriptorPool(ctx->device, MAX_FRAMES_IN_FLIGHT);
-    auto descriptor_set =
-        DescriptorSet(ctx, descriptor_set_layout, descriptor_pool, MAX_FRAMES_IN_FLIGHT,
-                      &uniform_buffers, &texture, &sampler);
-    geometry = std::make_unique<Geometry::Rectangle>(
-        ctx, swap_chain_manager, graphicsQueue, std::move(descriptor_set));
-}
-
-TextPipeline::~TextPipeline() {
-    // Order of these functions matter
-    vkDestroyDescriptorPool(ctx->device, descriptor_pool, nullptr);
-    vkDestroyPipeline(ctx->device, graphics_pipeline, nullptr);
-    vkDestroyPipelineLayout(ctx->device, pipeline_layout, nullptr);
+    vkDestroyShaderModule(m_ctx->device, vert_shader_module, nullptr);
+    vkDestroyShaderModule(m_ctx->device, frag_shader_module, nullptr);
+    return pipeline;
 }
 
 void TextPipeline::render_text(const VkCommandBuffer &command_buffer,
@@ -59,7 +57,7 @@ void TextPipeline::render_text(const VkCommandBuffer &command_buffer,
         std::forward<std::vector<StorageBufferObject>>(instance_data));
     const auto num_instances = instance_data.size();
     if (num_instances > 0) {
-        geometry->record_draw_command(command_buffer, graphics_pipeline, pipeline_layout,
-                                      num_instances);
+        geometry->record_draw_command(command_buffer, m_pipeline.m_pipeline,
+                                      m_pipeline.m_pipeline_layout, num_instances);
     }
 }
