+++ src/render_engine/TextPipeline.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "render_engine/CoreGraphicsContext.h"
+#include "render_engine/DescriptorPool.h"
+#include "render_engine/Pipeline.h"
 #include "render_engine/RenderableGeometry.h"
 #include "render_engine/Sampler.h"
 #include "render_engine/SwapChainManager.h"
@@ -17,6 +19,24 @@
 #include <glm/gtc/matrix_transform.hpp>
 
 class TextPipeline {
+  private:
+    std::shared_ptr<CoreGraphicsContext> m_ctx;
+
+    Pipeline m_pipeline;
+    DescriptorPool m_descriptor_pool;
+
+    std::unique_ptr<Geometry::Rectangle> geometry;
+
+    const std::vector<const char *> validationLayers = {"VK_LAYER_KHRONOS_validation"};
+
+    bool framebufferResized = false;
+    Pipeline create_pipeline(VkDescriptorSetLayout &descriptor_set_layout,
+                             SwapChainManager &swap_chain_manager);
+
+    bool checkDeviceExtensionSupport(const VkPhysicalDevice &physicalDevice);
+
+    void updateUniformBuffer(uint32_t currentImage);
+
   public:
     TextPipeline(Window &window, std::shared_ptr<CoreGraphicsContext> ctx,
                  SwapChainManager &swap_chain,
@@ -27,20 +47,4 @@ class TextPipeline {
 
     void render_text(const VkCommandBuffer &command_buffer,
                      std::vector<StorageBufferObject> &&instance_data);
-
-  private:
-    std::shared_ptr<CoreGraphicsContext> ctx;
-
-    VkPipelineLayout pipeline_layout;
-    VkPipeline graphics_pipeline;
-    VkDescriptorPool descriptor_pool;
-    std::unique_ptr<Geometry::Rectangle> geometry;
-
-    const std::vector<const char *> validationLayers = {"VK_LAYER_KHRONOS_validation"};
-
-    bool framebufferResized = false;
-
-    bool checkDeviceExtensionSupport(const VkPhysicalDevice &physicalDevice);
-
-    void updateUniformBuffer(uint32_t currentImage);
 };
